+++ packages/dash-table/src/dash-table/utils/generate.ts
@@ -1,0 +1,8 @@
+export default function genRandomId(prefix: string, radix: number = 36) {
+    return (
+        prefix +
+        Math.random()
+            .toString(radix)
+            .substring(2)
+    );
+}