+++ packages/dash-table/CHANGELOG.md
@@ -391,3 +391,6 @@ Derived properties allow the component to expose complex state that can be usefu
         dataframe_timestamp -> data_timestamp
         derived_virtual_dataframe -> derived_virtual_data
         derived_viewport_datafram -> derived_viewport_data
+
+## RC7 - Optional id prop
+    - The id prop of the table is now optional. It will generate a random id if it's not set.